+++ src/lib/config.ts
@@ -47,8 +47,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "4.17.2-alpha.1778022445373-0aa80216",
-  sdkVersion: "0.8.0-alpha.2",
-  genVersion: "2.881.17",
+  sdkVersion: "0.8.0-alpha.3",
+  genVersion: "2.882.0",
   userAgent:
-    "speakeasy-sdk/typescript 0.8.0-alpha.2 2.881.17 4.17.2-alpha.1778022445373-0aa80216 cribl-control-plane",
+    "speakeasy-sdk/typescript 0.8.0-alpha.3 2.882.0 4.17.2-alpha.1778022445373-0aa80216 cribl-control-plane",
 } as const;