+++ RELEASES.md
@@ -1559,3 +1559,13 @@ Based on:
 - [typescript v0.8.0-alpha.2] .
 ### Releases
 - [NPM v0.8.0-alpha.2] https://www.npmjs.com/package/cribl-control-plane/v/0.8.0-alpha.2 - .
+
+## 2026-05-08 16:47:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.8.0-alpha.3] .
+### Releases
+- [NPM v0.8.0-alpha.3] https://www.npmjs.com/package/cribl-control-plane/v/0.8.0-alpha.3 - .